// SDK/projects_api/client/models/TeamData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EntryHistory } from './EntryHistory';
import type { Pick } from './Pick';
export type TeamData = {
    active_chip?: (string | null);
    automatic_subs?: Array<any>;
    entry_history: EntryHistory;
    picks: Array<Pick>;
};